+++ test/examples/run_example.jl
@@ -5,7 +5,7 @@ using Ipopt
 using Cbc
 using Pavito
 
-include("circle.jl")
+include("nlp.jl")
 
 # MIP solver
 const cplex = optimizer_with_attributes(CPLEX.Optimizer, 
@@ -33,14 +33,13 @@ const alpine = optimizer_with_attributes(Alpine.Optimizer,
                                         # "minlp_solver" => juniper,
                                         "nlp_solver" => ipopt,  
                                         "mip_solver" => cplex,
-                                        "presolve_bt" => false,
+                                        "presolve_bt" => true,
                                         "presolve_bt_max_iter" => 5,
-                                        "disc_ratio" => 8,
-                                        "disc_consecutive_forbid" => true)
+                                        "disc_ratio" => 12)
 
 # Try different integer values ( >=4 ) for `disc_ratio` to have better Alpine run times 
 # Choose `presolve_bt` to `false` if you prefer the OBBT to be turned off. 
 
-m = circle(solver=alpine)
+m = nlp3(solver=alpine)
 
 JuMP.optimize!(m)